Remove unused asset server stuff
